refactor: Change import path to cell type.
